Add new configuration handling package
