+++ packages/config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ploys/deployments-core",
+  "name": "@ploys/deployments-config",
   "version": "1.0.0",
   "description": "Automated continuous deployment and delivery",
   "author": "Daniel Balcomb <daniel.balcomb@gmail.com>",
@@ -18,19 +18,15 @@
     "clean": "tsc --build tsconfig.build.json --clean"
   },
   "dependencies": {
-    "@octokit/auth-app": "^2.4.5",
+    "@hapi/joi": "^17.1.1",
     "@octokit/rest": "^17.9.2",
-    "@octokit/types": "^4.0.2",
-    "@octokit/webhooks": "^7.6.2",
     "await-to-js": "^2.1.1",
-    "eventsource": "^1.0.7"
+    "js-yaml": "^3.14.0"
   },
   "devDependencies": {
-    "@types/eventsource": "^1.1.2",
+    "@types/hapi__joi": "^17.1.0",
     "@types/jest": "^25.2.3",
     "@types/js-yaml": "^3.12.4",
-    "@types/node": "^14.0.5",
-    "js-yaml": "^3.14.0",
-    "nock": "^12.0.3"
+    "@types/node": "^14.0.5"
   }
 }
